Make the incompatibility message more explicit (#16359)
Discussion in nvda.groups.io/g/nvda/topic/clock_add_on_add_on_store/105313955?p=,,,100,0,0,0::recentpostdate/sticky,,,100,2,0,105313955,previd%3D1712173725527803825,nextid%3D1710680040743854194&previd=1712173725527803825&nextid=1710680040743854194

Summary of the issue:
The current incompatibility message for too old add-ons in the add-on store's details is as follows:

An updated version of this add-on is required. The minimum supported API version is now 2024.1. This add-on was last tested with 2023.1. You can enable this add-on at your own risk.

The concept of API is not clear for everyone and it is not explained in NVDA's GUI, nor in the User Guide. Thus, the logic of this sentence explaining why the add-on is not compatible is not easily understandable.

Description of user facing changes
Modify the incompatibility reason for too old add-ons as follows:

An updated version of this add-on is required. This add-on was last tested with 2023.1. The current version of NVDA requires that the extension be tested with NVDA 2024.1 or higher. You can enable this add-on at your own risk.