Create a new folder by default when making a portable copy, warn for non-empty directories (#16686)
Fixes #16443

Summary of the issue:
When creating a portable copy, NVDA overwrites the contents of the folder. This may delete files like dlls.
This caused #16514 to be reported, as this can be quite destructive and unexpected.

Typical behaviour for writing to a directory like this involves creating a new folder to write the contents to.
However, portable copies can also be upgraded, so it is important to be able to support writing to existing portable copy directories.

Description of user facing changes
When creating a portable copy, a new checkbox has been added to create a new folder for the portable copy, which is checked by default.
If this box is unchecked, a warning dialog will appear if the directory is non-empty.
The directory contains nvda.exe we assume its a portable copy directory, and confirm the user wishes to upgrade the portable copy.
The directory is otherwise non-empty, a general warning dialog is opened warning the user that the contents of the folder will be overwritten.