When handling caret movement, fetch the new caret info after checking for a pending caret event. (#16711)
Summary of the issue:
In Firefox's in-progress UIA implementation, moving with the cursor keys in editable text boxes sometimes reports the character at the old position of the caret instead of the new position of the caret. For example, on the line:
abcdef
with the cursor positioned on "a", pressing right arrow several times might read a, b, c instead of b, c, d.
I don't anticipate that NVDA will ever move to using UIA in Firefox. Nevertheless, this is an NVDA bug which might impact other things.

Description of user facing changes
When navigating with the cursor keys in text boxes in applications which use UI Automation, NVDA no longer sometimes reports the wrong character, word, etc.

Description of development approach
Previously, EditableText._hasCaretMoved fetched the caret before checking for a pending caret event. It then returned this fetched caret. The following race condition could occur:

In NVDA's main thread, NVDA queries the caret. The caret hasn't updated yet in the app, so this is the old position.
The app fires a caret event.
NVDA receives this caret event on a background UIA thread and queues it.
Back in NVDA's main thread, NVDA checks for a pending caret event. That check passes because of 3).
NVDA returns the old caret info retrieved in 1).
This can't happen with other APIs because they all queue events on the main thread, which can't be preempted during _hasCaretMoved. However, UIA queues events from a background thread.