Support for text review commands on object name in Visual Studio Code (#16248)
Fixed #16246

Summary of the issue:
Most controls in Visual Studio Code do not support text review commands.

The most typical example is when the user wants to review the control name character by character, such as the file name in the file tree and the setting items on the settings page. Based on current experience, this is not easy to achieve.

Description of user facing changes
Users will be able to review control name in Visual Studio Code by NVDA’s text review commands.

Description of development approach
In the application module for Visual Studio Code, a modification has been made within the event_NVDAObject_init method to enhance the handling of objects.
Specifically, for objects whose role is not EDITABLETEXT , the TextInfo property is now explicitly set to NVDAObjectTextInfo.