Fix remote ops to again allow access to the Windows Mail message area (#16797)
Fixes #16689

Summary of the issue:
The user can no longer read the Windows Mail message area with the arrow keys.
This is due to the fact that this implementation of the MS Word control does not support custom attribute values, but more importantly because code to fetch custom attribute values in NVDA via remote ops is not fail-safe enough.
In this case, the remote ops function was raising a NoReturnException as the function was not returning a value on common errors.

Description of user facing changes
The user can again read messages in Windows Mail.

Description of development approach
Rewrote the remote ops msWord_getCustomAttributeValue function to be more readable and return early on any error. this avoids deeply nested if statements, and ensures that all code paths return a value. This also required the following extra change:
UIA remote ops can now marshal the None object as a RemoteNull object. This None can now be returned from a remote ops function via RemoteAPI.Return.