Document incompatibility of Code Factory synthesizers with 'Use spelling functionality' (#16794)
Many threads on NVDA mailing list, the last ones being:

NVDA Vocalizer expressive spelling
To stop NVDA from spelling when not wanted, uncheck the setting "Use spelling functionality if
NVDA 2024-1 and Code Factory Vocaliser pronunciation problem
But there are much more of them.

Summary of the issue:
Code Factory synthesizers (NVDA add-on and SAPI one) are not compatible with "Use spelling functionality" setting. This needs to be documented. It's worth noting that the add-on is being discontinued, so there will never be a compatible version.

Description of user facing changes
See User Guide.

Description of development approach
I just mention Code Factory, not Vocalizer nor Eloquence, to avoid confusion with the other Vocalizer synthesizer from Tiflotecnia.