Improve ruff lint rules, indented functions and scons inbuilts (#16804)
Closes #11905
Follow up to #16767

Summary of the issue:
scons files have a lot of inbuilt functions that the linter doesn't recognize. See also #11905

Ruff preserves indent/new-line formatting of function arguments, list items, and similar iterables, if a trailing comma is added.
Much of NVDA is missing these trailing commas causing the lint to collapse them.

Description of user facing changes
Description of development approach
A lint rule has been added to ignore F821 errors in scons files.

A pre-commit hook is added, which adds a trailing comma to python in case it was missed.