Revert disabling handling of caret notifications for UIA (#16817)
Partially reverts #14888, #15838

Summary of the issue:
In #14888, XamlEditableText was added to ensure that UIA events wouldn't be used to determine caret changes, as the hypothesis was that they were fired too early. In #15838, I expanded this workaround for WPF (Visual Studio) text controls.
It turns out @jcsteh found the actual cause of these issues and fixed them in #16711, allowing us to rely on events again.

Description of user facing changes
None, though caret movement would possibly be detected a little bit faster in XAML and WPF controls.

Description of development approach
Mostly reverts.