Add a boolean to inform the architecture of the runing operating system (#16726)
Closes #16330

Summary of the issue:
As of 2024, Windows 10 is the last version to support 32-bit processors from Intel, AMD, and ARM, with Windows 11 supporintg 64-bit platforms only (AMD64 and ARM64). In 2025, consumer support for Windows 10 will end, with end of support for long-term servicing channel builds planned for 2032 (Windows 10 IoT Enterprise 2021 LTSC).
Originally, NVDA was written for 32-bit Python runtime. With impending end of support for 32-bit platforms, moving to 64-bit became a goal for NV Access and the NVDA community. According to #16304, a phased transition plan is proposed, with full transition tentatively scheduled for 2026.

Description of user facing changes
This way NVDA Core and add-ons can execute appropriate code paths. For example, when accessing Windows Registry, WoW64 paths must be used if running 32-bit NVDA on 64-bit Windows and vice versa.