Normalize path of crash dump (#16650)
When a crash dump is created, the following is logged:

CRITICAL - watchdog._crashHandler (14:32:14.383) - MainThread (15656):
NVDA crashed! Minidump written to C:\Users\Cyrille\Documents\DevP\GIT\nvda\source\nvda.log\..\nvda_crash.dmp
Although it causes no functional issue, the path is not normalized (presence of .. in the path). It is a bit confusing when reading this in the log.

Description of user facing changes
The message being logged shows a normalized path.

Description of development approach
Get the directory of the log file and join with the dumpfile file name.