Make getFormatFieldAtRange a little more concise (#16821)
N/A. This is purely a refactor and changes no behavior.

Summary of the issue:
UIATextInfo._getFormatFieldAtRange is flagged as being too complex (C901). This change makes it simpler.

Description of user facing changes
None

Description of development approach
Pull each of the if statements into a separate helper function. This makes _getFormatFieldAtRange much more concise and simple.

I am submitting this change as a separate PR because I plan to continue working on this function and want to separate the refactoring from the change in logic behavior. Also I changed a fair number of lines, so I want to get this in quickly to avoid future merge conflicts.