On-demand speech mode - Do not speak anymore during sayAll when SayAll is not launched by a SayAll script (#16826)
Fixes #16825

Summary of the issue:
While in on-demand speech mode, NVDA speaks during Say All. This is OK when Say All is launched by a specific Say All script, e.g. NVDA+downArrow. But this should not happen when Say All occurs in other situations, i.e. when a new page is loaded in a browser, when an e-mail is opened or when we go forward in a PowerPoint slideshow.

Description of user facing changes
In on-demand mode, NVDA will not talk anymore during Say All if Say All was not launched by an explicit Say All script.

Description of development approach
In speech/sayAll.py, add a new parameter to SayAllHandler.readObjects and to SayAllHandler.readText to determine if the Say All was started from an explicit Say All script or not. This parameter can be None if SayAll is resumed after a script's execution (skim reading case); in this case the previous value of startedFromScript is kept. This value is then used to know if the speech should be output or not.

While at it, I have added to on-demand scripts an Eclipse script to read the documentation; this script seems to have been forgotten during on-demand development.