Complete support of bk:space and bk:space+dots gestures for Seika Notetaker (#16828)
Closes #16642.

Summary of the issue:
Two drivers are mentioned in the issue, eurobraille and seikantk. The user can change an option of the former to solve the issue. However, there is no feasible approach for the user to solve the issue for the latter manually, which is the aim of the PR.

Description of user facing changes
The user can perceive the changes by three ways.

Observe the input help. On pressing dots and/or space/backspace, names of braille input gestures are reported correctly.
Run braille input translation. The result must be as expected, while the old seikantk driver behaves incorrectly.
Use add-ons relying on braille input. In fact, this is the main demand that users reqest for the modification. Correct braille input leads to correct result.
Description of development approach
4 cases of input gestures are explicitly specified in seikantk.py, and the algorithm description is available in the commit log.