UIA handler/notification event handler: attempt to fetch native window handle when the element reports nothing. Re #16871.
Sometimes a UIA element may say there is no native window handle when in fact it is part of its runtime ID. Therefore fetch the runtime ID tuple if possible, logging the process along the way. If the runtime ID tuple (really an array) is empty, return early as there is nothing NVDA can do at this time.