Remove carriage return and bell character from table header information in IAccessible MS Word (#16874)
Closes #7212

Summary of the issue:
When accessing Microsoft Word via IAccessible, column and row headers set with the set row/column header scripts in NVDA have '\r\a' at the end of them. While this does not seem to be an issue for speech (presumably synths just ignore control characters), these characters were being output in braille.

Description of user facing changes
Extraneous output is no longer produced.

Description of development approach
Modified NVDAObjects.IAccessible.winword.WordDocument.fetchAssociatedHeaderCellText to drop '\r\a' if it was a suffix to any of the row or column headers' text.