IA2TextTextInfo.expand: Correctly add the text start offset when adjusting for embedded objects during mouse navigation. (#16912)
In Firefox, when a paragraph, line, etc. contains links (or other elements), moving the mouse over text before or after the link sometimes doesn't report the text correctly. I attempted to fix this in #14755, and while my fix did work in some cases, it seems it wasn't quite complete.

Description of user facing changes
In Mozilla Firefox, moving the mouse over text before or after a link now reliably reports the text.

Description of development approach
To search for embedded objects, IA2TextTextInfo.expand fetches the text between the expanded start and end offsets. Previously, when adjusting the offsets if an embedded object was found, the code didn't add the offset from the start of the object to the start of the retrieved text. Now it does.