Catch and log exceptions from wasPlay_idle. (#16909)
Re #16125.

Summary of the issue:
IAudioClock::GetPosition, called from wasPlay_idle, sometimes fails with an access violation on a device which has been invalidated. This shouldn't happen and suggests a bug somewhere in NVDA's C++ WASAPI code. Currently, we keep calling this every 10 seconds, which just keeps logging the same error and isn't useful.

Description of user facing changes
When this (rarely) occurs, it won't keep logging errors and playing the error sound every 10 seconds. I haven't added a change log entry for this because it won't be noticeable to users on release, who normally don't have the error sound.

Description of development approach
I've tried to debug this issue, but thus far, I haven't been able to track it down far enough to fix it. In the meantime, the repeated error sounds and log spam are very annoying when this does occur. This should not be considered a fix for #16125; ideally, we need a proper fix for that. Even so, this will at least make it less painful. We still log the exception and play an error sound on the first instance so this can be caught and debugged further in future. We just don't keep doing it over and over again.