Indicate font attributes in braille with tags (#16864)
Closes #16863

Summary of the issue:
Braille users are currently limited in terms of the type of formatting attributes they can read in Braille: only bold, italics and underline are supported, and only if their chosen braille table supports the LibLouis emphasis codes for those attributes.

Description of user facing changes
Users can now configure NVDA to use starting and ending tags for bold, italics, underline and strike through. Users need to explicitly enable the new behaviour, the default is for font attributes to be displayed using native braille (if supported by the LibLouis table in use).

Description of development approach
Added a new FeatureFlagEnum that currently supports LibLouis and tags output modes, with the default being LibLouis. Added a control to the braille config to allow users to set their preferred display mode.

Updated braille.getFormatFieldBraille to generate a tag for formatting changes (via a private helper function). This new code is not called unless the user's preference is the new tags mode.

Updated braille.TextInfoRegion._getTypeformFromFormatField to return louis.plain_text if the user's preference is not to use LibLouis, in addition to the existing checks.