Add support for custom symbol dictionaries (#16823)
Closes #16739

Summary of the issue:
There is a desire to provide symbol pronunciation rules in optional symbol dictionaries, e.g. for ancient languages. IN #16739, we concluded that having the ability to provide them in add-ons would be most helpful.

Description of user facing changes
Removed the option Include Unicode Consortium data (including emoji) when processing characters and symbols from the speech settings category and replaced it by a checkable list box to enable optional dictionaries. This only contains the CLDR dictionary by default, but can be expanded by add-ons.
Added the ability for add-ons to provide extra dictionaries. They could be mandatory, meaning that enabling the add-on means enabling the dictionary. If optional, they are listed in the listbox mentioned above.
Description of development approach
Added a SymbolDictionaryDefinition class to characterProcessing that contains all data about a dictionary.
Changed the logic in characterProcessing to create speech symbol processors based on enabled definitions.
For dictionary registration in add-ons, I chose a similar approach as for custom braille tables, e.g. you define the dictionaries in a subsection of the add-on manifest.