Makes the timeout for multiple keypress configurable (#12049)
Fixes #11929.

Summary of the issue:
Some NVDA keystrokes can perform various commands according to the number of times the keystrokes is executed. E.g. NVDA+Numpad2 reports the current character under the review cursor with one key press. But a double key press reports the character description. And a tripe key press reports the character code.
Currently, double or triple key press should executed quite quickly to be considered a multiple key press: a maximum of 0.5 s is required between two key presses.
This may be difficult or even impossible for some people having physical disabilities and/or using sticky keys.

Description of user facing changes
This PR adds an option in the keyboard settings panel to allow to configure the timeout for multiple keypress.

The default value is the same as the current timeout, i.e. 500 ms.
The max value is 20000 s (= 20 s). I have taken a value that seemed quite exaggerated to me on purpose in order to be sure not to penalize anyone. I imagine that someone who is not able to type two keys in less than 20 seconds may not use a standard keyboard at all. But in any case, a real-world feedback is welcome.
Description of development approach
Added a value in ms in the GUI and in the config spec.