Handle "number of Braille cells" usage in HID Braille standard (#16916)
Standard HID Braille displays can inform the screen reader about the number of available Braille cells. For instance, an 80 cells Braille display can decide to use 20 cells for internal purposes and leave 60 cells for use by the screen reader. The device communicates this to the host by setting a certain HID usage to the number of available cells. The screen reader should still send as many cells as determined by the report descriptor, but all cells in excess of the number given in the "number of Braille cells" usage will be ignored by the device.

Description of user facing changes
The number of Braille cells assumed by NVDA will match the one given by the device in its "number of Braille cells" usage.

Description of development approach
The HID driver was modified to scan the report descriptor for the usage in question. If found, it will subsequently be h