Do not attempt to move out of bounds in IAccessible edit controls (#16927)
Closes #12885
Closes #16314

Summary of the issue:
When panning the display at the end of some edit controls, NVDA generates an error as it attempts to move to an offset that is out of range.

Description of user facing changes
The error is no longer encountered. This should not matter to most users, who would not have even known there was an error in the first place. As a result, I have not added a change log entry.

Description of development approach
Updated MozillaCompoundTextInfo.move to return 0 if the underlying object's text length is 0.
Added a check to IA2TextTextInfo.setCaretOffset to clamp offset to IAccessibleText::nCharacters and log a debug warning.