Add an --all-cores command line option to scons (#16943)
Fixup for #16868

Summary of the issue:
As part of #16868, i intended to implement support to pass a number of 0 cores to scons (e.g. scons source -j0) to automatically pick all available cores. It turns out that you cant override an option with SetOption when set on the command line.

Description of user facing changes
None, build system related.

Description of development approach
Added an --all-cores parameter that will pick all available cores.