Ensure that when character normalization results in a character with symbol definition, the symbol replacement is spoken (#16950)
Fixup for #16622

Summary of the issue:
When unicode normalization of a character (e.g. ·) resulted into a character that had a symbol definition (e.g. ·, middle dot), the symbol definition wasn't applied to the normalization. This resulted in NVDA speaking nothing or only the word normalized.

Description of user facing changes
NVDA will now properly speak the · character (Greek Ano Teleia) as middle dot when normalizing. This also applies to other characters where normalization results in a character that's part of the symbol dictionary.

Description of development approach
When normalizing a character, ensure it is thrown through characterProcessing.processSpeechSymbol.